backend/requirements.txt
asgiref==3.2.10
black==22.1.0
blue==0.9.1
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.0
django-colorfield==0.8.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
drf-extra-fields==3.4.1
exceptiongroup==1.1.1
filelock==3.12.0
flake8==4.0.1
fonttools==4.38.0
fpdf2==2.5.4
gunicorn==20.1.0
identify==2.5.24
idna==3.3
importlib-metadata
iniconfig==2.0.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.0
packaging==23.1
pathspec==0.11.1
Pillow==9.1.0
platformdirs==3.5.0
pluggy==1.0.0
progress==1.6
psycopg2-binary==2.8.6
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pytest==7.3.1
pytest-django==4.5.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2020.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.3.1
testfixtures==6.18.5
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.15.0
